+++ src/app/notes/trash/page.tsx
@@ -6,13 +6,10 @@ import { Loader2 } from 'lucide-react';
 import NotesRender from '@/components/shared/NotesRender';
 import NotesHeader from '@/components/business/market/NotesHeader';
 import SidebarMenu from '@/components/menu/sidebarMenu';
-import {
-  menuItemsBottom,
-  menuItemsTop,
-  notesMenu,
-} from '@/config/menuItems/business/dashboardMenuItems';
+import { notesMenu } from '@/config/menuItems/business/dashboardMenuItems';
 import useFetchNotes from '@/hooks/useFetchNotes';
 import Header from '@/components/header/header';
+import { menuItemsBottom } from '@/config/menuItems/freelancer/dashboardMenuItems';
 
 const TrashPage = () => {
   const userId = useSelector((state: any) => state.user?.uid);
@@ -34,7 +31,7 @@ const TrashPage = () => {
       <div className="flex flex-col sm:gap-8 sm:py-0 sm:pl-14 mb-8">
         <div>
           <Header
-            menuItemsTop={menuItemsTop}
+            menuItemsTop={notesMenu}
             menuItemsBottom={menuItemsBottom}
             activeMenu="Trash"
             breadcrumbItems={[
