+++ src/config/menuItems/business/settingsMenuItems.tsx
@@ -1,4 +1,4 @@
-import { HomeIcon, User } from 'lucide-react';
+import { HomeIcon, User, Verified } from 'lucide-react';
 import Image from 'next/image';
 
 import { MenuItem } from '@/components/menu/sidebarMenu';
@@ -23,9 +23,14 @@ export const menuItemsTop: MenuItem[] = [
     label: 'Home',
   },
   {
-    href: '/settings/business-info',
+    href: '/business/settings/business-info',
     icon: <User className="h-5 w-5" />,
     label: 'Business Info',
+  },
+  {
+    href: '/business/settings/kyc',
+    icon: <Verified className="h-5 w-5" />,
+    label: 'kyc',
   },
 ];
